Use init_control in test_cpu_* tests

+++ integration/test/test_cpu_activity_agent.py
@@ -31,7 +31,6 @@ from apps.arithmetic_intensity import arithmetic_intensity
 from apps.minife import minife
 
 @util.skip_unless_config_enable('beta')
-@util.skip_unless_do_launch()
 @util.skip_unless_workload_exists("apps/arithmetic_intensity/ARITHMETIC_INTENSITY/bench_sse")
 @util.skip_unless_workload_exists("apps/arithmetic_intensity/ARITHMETIC_INTENSITY/bench_avx2")
 @util.skip_unless_workload_exists("apps/arithmetic_intensity/ARITHMETIC_INTENSITY/bench_avx512")
@@ -42,18 +41,18 @@ class TestIntegration_cpu_activity(unittest.TestCase):
         """
         Setup applications, execute, and set up class variables.
         """
+        cls._skip_launch = not util.do_launch()
 
-        # Enable QM to measure total memory bandwidth
-
-        # Assign all cores to resource monitoring association ID 0. This
-        # allows for monitoring the resource usage of all cores.
-        geopm_test_launcher.geopmwrite("MSR::PQR_ASSOC:RMID board 0 {}".format(0))
-        # Assign the resource monitoring ID for QM Events to match the per
-        # core resource association ID above (0)
-        geopm_test_launcher.geopmwrite("MSR::QM_EVTSEL:RMID board 0 {}".format(0))
-        # Select monitoring event ID 0x2 - Total Memory Bandwidth Monitoring.
-        # This is used to determine the Xeon Uncore utilization.
-        geopm_test_launcher.geopmwrite("MSR::QM_EVTSEL:EVENT_ID board 0 {}".format(2))
+        # Define global init control config
+        #   Enable QM to measure total memory bandwidth for uncore utilization
+        mem_bw_cfg = """
+        # Assign all cores to resource monitoring association ID 0
+        MSR::PQR_ASSOC:RMID board 0 0
+        # Assign the resource monitoring ID for QM Events to match ID 0
+        MSR::QM_EVTSEL:RMID board 0 0
+        # Select monitoring event ID 0x2 - Total Memory Bandwidth Monitoring
+        MSR::QM_EVTSEL:EVENT_ID board 0 2
+        """
 
         # Grabbing system frequency parameters for experiment frequency bounds
         cpu_base_freq = geopm_test_launcher.geopmread("CPU_FREQUENCY_STICKER board 0")
@@ -67,20 +66,30 @@ class TestIntegration_cpu_activity(unittest.TestCase):
 
         mach = machine.init_output_dir('.')
 
-        def launch_helper(experiment_type, experiment_args, app_conf, experiment_cli_args):
-            output_dir = experiment_args.output_dir
-            if output_dir.exists() and output_dir.is_dir():
-                shutil.rmtree(output_dir)
+        cls._run_count = 0
+        def launch_helper(experiment_type, experiment_args, app_conf, experiment_cli_args, init_control_cfg):
+            if not cls._skip_launch:
+                init_control_path = 'init_control_{}'.format(cls._run_count)
+                cls._run_count += 1
+                with open(init_control_path, 'w') as outfile:
+                    outfile.write(init_control_cfg)
+                experiment_args.init_control = os.path.realpath(init_control_path)
 
-            experiment_type.launch(app_conf=app_conf, args=experiment_args,
-                                   experiment_cli_args=experiment_cli_args)
+                output_dir = experiment_args.output_dir
+                if output_dir.exists() and output_dir.is_dir():
+                    shutil.rmtree(output_dir)
+
+                experiment_type.launch(app_conf=app_conf, args=experiment_args,
+                                       experiment_cli_args=experiment_cli_args)
 
         ################
         # Monitor Runs #
         ################
-        geopm_test_launcher.geopmwrite("CPU_FREQUENCY_MAX_CONTROL board 0 {}".format(cpu_max_freq))
-        geopm_test_launcher.geopmwrite("CPU_UNCORE_FREQUENCY_MAX_CONTROL board 0 {}".format(uncore_max_freq))
-        geopm_test_launcher.geopmwrite("CPU_UNCORE_FREQUENCY_MIN_CONTROL board 0 {}".format(uncore_min_freq))
+        freq_cfg = """
+        CPU_FREQUENCY_MAX_CONTROL board 0 {}
+        CPU_UNCORE_FREQUENCY_MAX_CONTROL board 0 {}
+        CPU_UNCORE_FREQUENCY_MIN_CONTROL board 0 {}
+        """.format(cpu_max_freq, uncore_max_freq, uncore_min_freq)
 
         # MiniFE
         cls._minife_monitor_dir = Path(os.path.join('test_cpu_activity_output', 'minife_monitor'))
@@ -94,7 +103,7 @@ class TestIntegration_cpu_activity(unittest.TestCase):
             verbose=False,
         )
 
-        launch_helper(monitor, experiment_args, minife.MinifeAppConf(node_count), [])
+        launch_helper(monitor, experiment_args, minife.MinifeAppConf(node_count), [], mem_bw_cfg + freq_cfg)
 
         # Arithmetic Intensity Benchmark
         cls._aib_monitor_dir = Path(os.path.join('test_cpu_activity_output', 'aib_monitor'))
@@ -113,7 +122,7 @@ class TestIntegration_cpu_activity(unittest.TestCase):
         experiment_args.output_dir = cls._aib_monitor_dir
         experiment_args.trial_count = 3
 
-        launch_helper(monitor, experiment_args, aib_app_conf, [])
+        launch_helper(monitor, experiment_args, aib_app_conf, [], mem_bw_cfg + freq_cfg)
 
         ##################################
         # Uncore frequency sweep at base #
@@ -132,7 +141,7 @@ class TestIntegration_cpu_activity(unittest.TestCase):
         experiment_cli_args=['--geopm-report-signals={}'.format(report_signals)]
 
         # We're reusing the AIB app conf from above here
-        launch_helper(uncore_frequency_sweep, experiment_args, aib_app_conf, experiment_cli_args)
+        launch_helper(uncore_frequency_sweep, experiment_args, aib_app_conf, experiment_cli_args, mem_bw_cfg)
 
         ##############
         # Parse data #
@@ -155,7 +164,7 @@ class TestIntegration_cpu_activity(unittest.TestCase):
         experiment_args.min_uncore_frequency = uncore_efficient_freq
         experiment_args.max_uncore_frequency = uncore_efficient_freq
 
-        launch_helper(uncore_frequency_sweep, experiment_args, aib_app_conf, experiment_cli_args)
+        launch_helper(uncore_frequency_sweep, experiment_args, aib_app_conf, experiment_cli_args, mem_bw_cfg)
 
         ##############
         # Parse data #
@@ -181,10 +190,6 @@ class TestIntegration_cpu_activity(unittest.TestCase):
         ################################
         # CPU Activity Agent phi sweep #
         ################################
-        geopm_test_launcher.geopmwrite("CPU_FREQUENCY_MAX_CONTROL board 0 {}".format(cpu_max_freq))
-        geopm_test_launcher.geopmwrite("CPU_UNCORE_FREQUENCY_MAX_CONTROL board 0 {}".format(uncore_max_freq))
-        geopm_test_launcher.geopmwrite("CPU_UNCORE_FREQUENCY_MIN_CONTROL board 0 {}".format(uncore_min_freq))
-
         # Arithmetic Intensity Benchmark
         cls._aib_agent_dir = Path(os.path.join('test_cpu_activity_output', 'aib_cpu_activity'))
 
@@ -199,14 +204,14 @@ class TestIntegration_cpu_activity(unittest.TestCase):
             phi_list=[0.2, 0.5, 0.7],
         )
 
-        launch_helper(cpu_activity, ca_experiment_args, aib_app_conf, [])
+        launch_helper(cpu_activity, ca_experiment_args, aib_app_conf, [], mem_bw_cfg + freq_cfg)
 
         # MiniFE
         cls._minife_agent_dir = Path(os.path.join('test_cpu_activity_output', 'minife_cpu_activity'))
         ca_experiment_args.output_dir = cls._minife_agent_dir
         ca_experiment_args.trial_count = 1
 
-        launch_helper(cpu_activity, ca_experiment_args, minife.MinifeAppConf(node_count), [])
+        launch_helper(cpu_activity, ca_experiment_args, minife.MinifeAppConf(node_count), [], mem_bw_cfg + freq_cfg)
 
     def tearDown(self):
         if sys.exc_info() != (None, None, None):
